pyproject.toml
[tool.poetry]
name = "lyss-chat-gpt-tools"
version = "0.1.0"
description = ""
authors = ["MoonLyss <shaitan.lyssm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.3.5"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
